+++ lib/src/run_it.dart
@@ -9,36 +9,50 @@ extension ExtOnType<T> on T {
   ///
   /// Good to use on data that have same Type.
   /// ```
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   T runOn(T Function(T it) fn) => fn.call(this);
 
   /// Run a function on a value of any type without any required return.
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   void let(void Function(T it) fn) => fn.call(this);
 
-  /// Return I and receive T.
+  /// Return R and receive T.
   /// Ex:
   /// ```dart
   /// bool isTrue = true;
   /// String value = isTrue.run((it) => it ? 'true' : 'false');
   ///
-  /// // I : String : value : will return String
+  /// // R : String : value : will return String
   /// // T : bool : isTrue : will receive bool (it)
   /// ````
-  I run<I>(I Function(T it) fn) => fn.call(this);
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
+  R run<R>(R Function(T it) fn) => fn.call(this);
 
   /// Check whether the value is null or not.
   ///
   /// > No favorable to use on dynamic type.
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool get isNull => this == null;
+}
 
+extension ExtOnTypeNullCheckers<T> on T {
   /// Check if provided value is not null.
   ///
   /// > No favorable to use on dynamic type.
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
   bool get isNotNull => this != null;
 
-  /// Run callback if value is not null and return [I] or [I?]
+  /// Run callback if value is not null and return [R] or [R?]
   ///
   /// > No favorable to use on dynamic type.
-  I? ifNotNull<I>(I Function(T it) fn) => isNotNull ? fn.call(this) : null;
+  @pragma('vm:prefer-inline')
+  @pragma('dart2js:tryInline')
+  R? ifNotNull<R>(R Function(T it) fn) => isNotNull ? fn.call(this) : null;
 }
 
 /// Run a code at any where and return the expected Type.
@@ -49,4 +63,6 @@ extension ExtOnType<T> on T {
 /// ```
 ///
 /// return type are otpional.
+@pragma('vm:prefer-inline')
+@pragma('dart2js:tryInline')
 T run<T>(T Function() fn) => fn.call();