+++ CHANGELOG.md
@@ -5,3 +5,8 @@
 ## 1.0.1
 
 - add Git Repository.
+
+# 1.1.1
+
+- add @pragma('vm:prefer-inline') & @pragma('dart2js:tryInline') on each method.
+- separate null check extensions to own scope.